fmt(exec): separate file
