+++ includes/minishell.h
@@ -24,7 +24,9 @@ void	debug_func(const char *func_name, const int line_num);
 void	debug_2d_array(char **array);
 
 /* exec */
+void	child_process(char **commands, char **environ);
 int		execute_command(char **exec_command);
+int		parent_process(int *last_exit_status);
 
 /* input */
 char	*input_line(void);
